fix logger handling
