+++ examples/middleware/main.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"github.com/sillen102/simba"
-	"github.com/sillen102/simba/logging"
 )
 
 type ResponseBody struct {
@@ -34,6 +33,6 @@ func main() {
 	})
 	app.Router.POST("/users", simba.HandlerFunc(handler))
 
-	logging.Get().Info().Msg("Listening on http://localhost:9999")
+	app.GetLogger().Info().Msg("Listening on http://localhost:9999")
 	http.ListenAndServe(":9999", app)
 }
