+++ examples/simple/main.go
@@ -6,7 +6,6 @@ import (
 	"net/http"
 
 	"github.com/sillen102/simba"
-	"github.com/sillen102/simba/logging"
 )
 
 type RequestBody struct {
@@ -49,6 +48,6 @@ func main() {
 	app.Router.POST("/users", simba.HandlerFunc(handler))
 	app.Router.GET("/no-body", simba.HandlerFunc(noBodyHandler))
 
-	logging.Get().Info().Msg("Listening on http://localhost:9999")
+	app.GetLogger().Info().Msg("Listening on http://localhost:9999")
 	http.ListenAndServe(":9999", app)
 }
